Multitap: ext clock quantizes taps when Sync is off
Quantize is on when Ext Clock < 0.5Hz and Sync is off.
In this case, all new taps are shifted to the closest multiple of the Ext Clock period.
Quantize turns off when Ext Clock > 0.5Hz or when Sync is turned on.
Quantize status is indicated by a blue DEL LED.